cloudclient/aws: Move utilities from machine_helper.go
